Move usage to separate package
